requirements.txt
Django>=5.1.4
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
Pillow>=12.1.0
python-decouple>=3.8

